requirements.txt
python-dotenv
fastapi
pydantic
requests
uvicorn